Tests for event downlink.
